backend/app/engine/pyproject.toml
[project]
name = "zenithalgo"
version = "0.1.2"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.8",
    "pandas>=2.3.3",
    "pytest>=9.0.2",
    "PyYAML>=6.0",
    "requests>=2.31",
    "seaborn>=0.13.2",
    "websockets>=11.0",
    "pydantic>=2.0.0",
    "pyarrow>=15.0.0",
    "plotly>=6.5.0",
    "jinja2>=3.1.6",
    "psycopg2-binary>=2.9.11",
    "sqlalchemy>=2.0.45",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
markers = [
    "live: mark a test as requiring live exchange/network",
]

[build-system]
requires = ["maturin>=1.0,<2.0"]
build-backend = "maturin"

[tool.maturin]
manifest-path = "../../native/core/Cargo.toml"
features = ["pyo3/extension-module"]


